backend/requirements.txt
flask==3.0.2
flask-cors==5.0.1
spacy==3.8.4
scikit-learn==1.6.1
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.4

# Additional dependencies
numpy==1.26.4
pandas==2.2.1
